pyproject.toml
[tool.poetry]
name = "fastapi-observability-example"
version = "1.1.3"
description = "fastapi-observability-example"
authors = ["Thibault Ayanides <thibault.serti@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
opentelemetry-api = "^1.21.0"
opentelemetry-sdk = "^1.21.0"
redis = "^5.0.1"
prometheus-fastapi-instrumentator = "^6.1.0"
opentelemetry-instrumentation-fastapi = "^0.42b0"
opentelemetry-instrumentation-redis = "^0.42b0"
httpx = "^0.25.1"
opentelemetry-exporter-otlp = "^1.21.0"
pyroscope-io = "^0.8.5"


[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.8"
black = "^23.1.0"
pylint = "^2.17.0"
mypy = "^1.1.1"
isort = "^5.12.0"
pytest = "^7.2.2"
coverage = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[tool.isort]
profile = "black"
force_single_line = false
atomic = false
include_trailing_comma = true
lines_after_imports = 2
lines_between_types = 1
use_parentheses = true
filter_files = true
known_first_party = "poetry"

[tool.mypy]
ignore_missing_imports = true


[tool.pylint.messages_control]
max-line-length = 120

extension-pkg-whitelist = "pydantic"

disable = [
]
